pyproject.toml
[tool.poetry]
name = "django-user-menagment"
version = "0.1.0"
description = ""
authors = ["Minister944 <kacper.machel.km@gmail.com>"]
readme = "README.md"
packages = [{include = "django_user_menagment"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
crispy-bootstrap5 = "^0.7"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-cors-headers = "^3.13.0"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"
isort = "^5.12.0"
freezegun = "^1.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"